Free some SRAM by putting some code in flash
Modified the linker script to put code into flash. Created
program_flash.cpp to move code that must be run from SRAM out of
ZuluSCSI_platform.cpp. Now ZuluSCSI_platform.cpp is running from flash.
Also running the following out of flash
- ZuluSCSI_log.cpp
- ZuluSCSI_log_trace.cpp
- ZuluSCSI_settings.cpp
- QuirksCheck.cpp
