+++ lib/ZuluSCSI_platform_RP2040/rp2040-daynaport.ld
@@ -73,6 +73,11 @@ SECTIONS
         /* Put only non-timecritical code in flash
          * This includes e.g. floating point math routines.
          */
+        .pio/build/ZuluSCSI_Pico_DaynaPORT/src/ZuluSCSI_log.cpp.o(.text .text*)
+        .pio/build/ZuluSCSI_Pico_DaynaPORT/src/ZuluSCSI_log_trace.cpp.o(.text .text*)
+        .pio/build/ZuluSCSI_Pico_DaynaPORT/src/ZuluSCSI_settings.cpp.o(.text .text*)
+        .pio/build/ZuluSCSI_Pico_DaynaPORT/src/QuirksCheck.cpp.o(.text .text*)
+        *libZuluSCSI_platform_RP2040.a:ZuluSCSI_platform.cpp.o(.text .text*)
         *libm*:(.text .text*)
         *libc*:(.text .text*)
         *libgcc*:*df*(.text .text*)
