+++ platformio.ini
@@ -112,6 +112,7 @@ build_flags =
     -DHAS_SDIO_CLASS
     -DUSE_ARDUINO=1
     -DZULUSCSI_V2_0
+    -Wl,-Map=${BUILD_DIR}/${PROGNAME}.map
 
 ; ZuluSCSI RP2040 hardware platform, as above, but with audio output support enabled
 [env:ZuluSCSI_RP2040_Audio]
